COUNTER_Robots_list.json: Remove unnecessary escaping of dashes
It is not necessary to escape dashes (-) unless they are used in a
character class where they could otherwise be used to note a range.

For example, in the expression "[a-z]" the dash indicates a range
of characters from a to z. If you wanted to match a literal dash in
a character class you would have to use "[a-z\-]", meaning a to z
or a dash. As none of our uses are inside character classes, we do
not need to escape them as literals.

I verified this by testing Java's java.util.regex.Pattern as well
as several PCRE implementations.

See: https://www.freeformatter.com/java-regex-tester.html
See: https://regex101.com/