COUNTER_Robots_list.json: Add Pattern
Pattern is a web mining module for the Python programming language.

---

This seems to be an academic spider. The spider's user agent looks
like this in my logs:

    Pattern/2.6 +http://www.clips.ua.ac.be/pattern

Because the word "Pattern" is not very unique and could appear in
some legitimate human user's user agent I suggest using anchoring
the string to the beginning of the line and matching at least one
digit for the version.

See: https://www.clips.uantwerpen.be/pattern