COUNTER_Robots_list.json: Add OgScrper
This user agent is responsible for tens of thousands of hits to my
web server over the last few years:

    OgScrper/1.0.0

There is very little information about this client on the web, but
it is included in several other robot detection libraries. As the
user agent is sufficiently unique I don't think we need to worry
about matching the version number.