COUNTER_Robots_list.json: Escape literal dots
Some patterns are incorrect because they mean to use a literal dot,
but we should technically be escaping these to use literal dots in
regex. Note that we need to escape the first backslash in the JSON
so we can get a literal backslash for escaping the dots in the out-
put file. The dot in the "virus.detector" pattern seems to be inte-
ntional, as this user agent sometimes appears as virus-detector and
other times as virus_detector.

Also I've taken the liberty to add URLs and descriptions where app-
licable.